requirements.txt
streamlit
paho-mqtt
plotly
